+++ packages/types/lib/scanner.d.ts
@@ -159,12 +159,12 @@ export interface Invasion {
   grunt_type: number
   incident_expire_timestamp: number
   confirmed: boolean
-  slot_1_pokemon_id: number
-  slot_1_form: number
-  slot_2_pokemon_id: number
-  slot_2_form: number
-  slot_3_pokemon_id: number
-  slot_3_form: number
+  slot_1_pokemon_id: number | null
+  slot_1_form: number | null
+  slot_2_pokemon_id: number | null
+  slot_2_form: number | null
+  slot_3_pokemon_id: number | null
+  slot_3_form: number | null
 }
 
 export interface ShowcaseEntry extends PokemonDisplay {