+++ src/hooks/useTranslateById.js
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next'
 import { getFormDisplay } from '@utils/getFormDisplay'
 
 /**
- * @typedef {{ plural?: boolean, amount?: boolean, alt?: boolean, newLine?: boolean, quest?: boolean, omitFormSuffix?: boolean }} CustomTOptions
+ * @typedef {{ plural?: boolean, amount?: boolean, alt?: boolean, newLine?: boolean, quest?: boolean, omitFormSuffix?: boolean, showDefaultForms?: boolean }} CustomTOptions
  * @typedef {(id: string, options?: CustomTOptions) => string} CustomT
  */
 
@@ -21,7 +21,15 @@ export function useTranslateById(options = {}) {
       language: i18n.i18n.language,
       t: (
         id,
-        { plural, amount, alt, newLine, quest, omitFormSuffix } = options,
+        {
+          plural,
+          amount,
+          alt,
+          newLine,
+          quest,
+          omitFormSuffix,
+          showDefaultForms,
+        } = options,
       ) => {
         if (typeof id !== 'string') {
           return ''
@@ -118,7 +126,7 @@ export function useTranslateById(options = {}) {
               form === undefined
                 ? ''
                 : getFormDisplay(pokemon, form, undefined, {
-                    showDefaultForms: quest,
+                    showDefaultForms: showDefaultForms || quest,
                     appendFormSuffix: !omitFormSuffix,
                   })
             return formLabel
@@ -134,7 +142,9 @@ export function useTranslateById(options = {}) {
       options.amount,
       options.plural,
       options.newLine,
+      options.quest,
       options.omitFormSuffix,
+      options.showDefaultForms,
     ],
   )
 }